[core] Remove `gcs_rpc_server.h` dependency from `RuntimeEnvHandler` (#55901)
Depends on: https://github.com/ray-project/ray/pull/55886

---------
