+++ src/ray/gcs/gcs_server/grpc_service_interfaces.h
@@ -64,5 +64,14 @@ class NodeInfoGcsServiceHandler {
                                     SendReplyCallback send_reply_callback) = 0;
 };
 
+class RuntimeEnvGcsServiceHandler {
+ public:
+  virtual ~RuntimeEnvGcsServiceHandler() = default;
+
+  virtual void HandlePinRuntimeEnvURI(PinRuntimeEnvURIRequest request,
+                                      PinRuntimeEnvURIReply *reply,
+                                      SendReplyCallback send_reply_callback) = 0;
+};
+
 }  // namespace rpc
 }  // namespace ray
