+++ src/ray/gcs/gcs_server/grpc_services.cc
@@ -36,5 +36,13 @@ void NodeInfoGrpcService::InitServerCallFactories(
   RPC_SERVICE_HANDLER(NodeInfoGcsService, CheckAlive, max_active_rpcs_per_handler_)
 }
 
+void RuntimeEnvGrpcService::InitServerCallFactories(
+    const std::unique_ptr<grpc::ServerCompletionQueue> &cq,
+    std::vector<std::unique_ptr<ServerCallFactory>> *server_call_factories,
+    const ClusterID &cluster_id) {
+  RPC_SERVICE_HANDLER(
+      RuntimeEnvGcsService, PinRuntimeEnvURI, max_active_rpcs_per_handler_);
+}
+
 }  // namespace rpc
 }  // namespace ray
