+++ src/ray/gcs/gcs_server/runtime_env_handler.h
@@ -13,11 +13,14 @@
 // limitations under the License.
 
 #pragma once
+#include <boost/asio.hpp>
 #include <memory>
 #include <utility>
 
+#include "ray/common/asio/instrumented_io_context.h"
 #include "ray/common/runtime_env_manager.h"
-#include "ray/rpc/gcs/gcs_rpc_server.h"
+#include "ray/gcs/gcs_server/grpc_service_interfaces.h"
+
 namespace ray {
 namespace gcs {
 
@@ -25,7 +28,7 @@ typedef std::function<std::shared_ptr<boost::asio::deadline_timer>(std::function
                                                                    uint32_t delay_ms)>
     DelayExecutorFn;
 
-class RuntimeEnvHandler : public rpc::RuntimeEnvHandler {
+class RuntimeEnvHandler : public rpc::RuntimeEnvGcsServiceHandler {
  public:
   RuntimeEnvHandler(instrumented_io_context &io_service,
                     RuntimeEnvManager &runtime_env_manager,
